review: move Roadmap to Model
