+++ app/Blocks/roadmap/render.php
@@ -3,7 +3,8 @@
  * @see https://github.com/WordPress/gutenberg/blob/trunk/docs/reference-guides/block-api/block-metadata.md#render
  */
 
-use EasyRoadmap\Controller\Common\Roadmap;
+use EasyRoadmap\Model\Roadmap;
+
 
 //TODO: add product attribute
 //echo get_block_wrapper_attributes();
